fix sections report+diagram
