+++ src/components/Report/Report/Report.module.scss
@@ -1,3 +1,29 @@
+.section {
+  display: flex;
+  flex-wrap: wrap;
+  margin-left: auto;
+  margin-right: auto;
+  //   width: 100%;
+  align-content: center;
+  justify-content: center;
+  align-items: center;
+  background-color: #ffffff;
+  @media screen and (min-width: 768px) {
+    background-color: #ffffff;
+    box-shadow: 0px 10px 60px rgba(170, 178, 197, 0.2);
+    border-radius: 30px;
+    margin-bottom: 30px;
+    // padding: 20px;
+    padding-top: 20px;
+    margin-left: 50px;
+    margin-right: 50px;
+  }
+  @media screen and (min-width: 1280px) {
+    max-width: 1060px;
+    margin-left: 110px;
+    margin-right: 110px;
+  }
+}
 .tab {
   display: flex;
   justify-content: center;
@@ -9,6 +35,7 @@
   display: flex;
   flex-direction: row;
   flex-wrap: nowrap;
+
   padding: 5px 25px;
   color: #000000;
   font-weight: 700;
