+++ src/components/StatisticDiagram/Diagram.module.css
@@ -6,22 +6,29 @@
 .bar__accent {
   fill: var(--accent-color);
 }
-@media screen and (min-width: 768px){
-.container {
-  width: 100%;
-  margin-bottom: 76px;
-  padding: 15px 20px 15px 20px;
-  background: #fff;
+@media screen and (min-width: 768px) {
+  .container {
+    /* width: 100%; */
+    margin-bottom: 76px;
+    padding: 15px 20px 15px 20px;
 
-  border-radius: 30px;
-  box-shadow: var(--box-shadow-tablet-desktop);
-}
+    padding-top: 20px;
+    margin-left: 50px;
+    margin-right: 50px;
+    background: #fff;
+
+    border-radius: 30px;
+    box-shadow: var(--box-shadow-tablet-desktop);
+  }
 }
 
 @media screen and (min-width: 1280px) {
   .container {
-    width: 100%;
+    /* width: 100%; */
     margin-bottom: 74px;
+    max-width: 1060px;
+    margin-left: 110px;
+    margin-right: 110px;
     padding: 20px 151px 20px 151px;
   }
 }
